+++ .gitignore
@@ -3,6 +3,8 @@ logs
 *.log
 npm-debug.log*
 
+dist
+
 package-lock.json
 
 # Runtime data
