+++ auth/AuthService.js
@@ -7,10 +7,13 @@ const isValidUser = async ({ username, password:textPassword }) => {
   try {
     // find the user by username
     const user = await User.findOne(username);
+    console.log(user)
 
     // compare the cleartext password and the pwd from the db
     // using bcrypt
-    //return await bcrypt.compare(textPassword, user.password);
+    const goodPassword = await bcrypt.compare(textPassword, user.password)
+    console.log({goodPassword});
+    return goodPassword;
   } catch (err) {
     console.error(err);
     return false;
@@ -18,29 +21,37 @@ const isValidUser = async ({ username, password:textPassword }) => {
 };
 
 module.exports = {
-  async authenticate(req, res, next) {
-    if (await isValidUser(req.body)) {
-      res.locals.token = await TokenService.makeToken({
-        username: req.body.username,
-        roles: ['admin', 'editor'],
-      });
-      return next();
+  authenticate(req, res, next) {
+    if (!isValidUser(req.body)) {
+      return next({});
     }
 
-    return next({});
+    TokenService.makeToken({
+        username: req.body.username,
+        roles: ['admin', 'editor'],
+      }).then((token) => {
+         res.locals.token = token;
+         next();
+      }).catch((err) => {
+        debugger;
+        next(err);
+      });
+      return false
   },
 
-  isAuth(roles) {
+  allow({roles}) {
     return [
-      async function isAuth(req, res, next) {
-        const token = req.headers.authorization;
-        const isValid = await TokenService.verify(token)
-        debugger;
+      (req, res, next) => {
+        TokenService.verify(req.authToken)
+        .then((payload) => {
+          const isAuthorized = roles.some (n => payload.roles.includes(n));
+          return isAuthorized ? next() : Promise.reject('User not authorized');
+        })
+        .catch(next);
       },
       (err, req, res, next) => res.status(403).json({})
     ];
-  },
-
+  }
 };
