+++ server.js
@@ -1,9 +1,11 @@
+require('dotenv').config();
 const express = require('express');
 const path = require('path');
 const logger = require('morgan');
 const bodyParser = require('body-parser');
 const cors = require('cors');
 const { authRouter, AuthService } = require('./auth');
+const tokenService = require('./auth/TokenService');
 
 const app = express();
 
@@ -20,6 +22,7 @@ app.use(cors());
 app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: false}));
+app.use(tokenService.receiveToken);
 
 app.get('/', function(req, res) {
   res.sendFile(__dirname + '/public/index.html');
@@ -28,10 +31,10 @@ app.get('/', function(req, res) {
 //ROUTE HANDLER
 const scoresRouter = require('./routes/scores');
 app.use('/auth', authRouter);
-app.use('/api/scores', AuthService.isAuth(), scoresRouter);
+app.use('/api/scores', AuthService.allow({roles: []}), scoresRouter);
 
 //const scoresRouter = require('./routes/scores');
-app.use('/api/scores', scoresRouter);
+//app.use('/api/scores', scoresRouter);
 
 app.get('*', function(req, res) {
   res.status(404).send({message: 'Hmm...Not Found.'});
