+++ src/calculateResult.js
