+++ src/components/SingleScore.js
@@ -56,7 +56,7 @@ class SingleScore extends Component {
   handleCancel() {
     this.setState({
       isEditing: false
-    });
+    })
   }
 
   handleSubmit(event) {
@@ -83,19 +83,18 @@ class SingleScore extends Component {
   renderScore() {
     if (this.state.scoreDataReceived === true) {
     return (
-      <div>
-      <form onSubmit={this.handleSubmit}>
+    <div>
+     <form onSubmit={this.handleSubmit}>
       <div className="my-score">
-      <div id="imgList">
-      <img src={this.state.score.url}/>
-      </div>
-      <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="happy" type="text" value={this.state.score.happy}/> : this.state.score.happy}</h3>
-      <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="mad" type="text" value={this.state.score.mad}/> : this.state.score.mad}</h3>
-      <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="url" type="text" value={this.state.score.url}/> : this.state.score.url}</h3>
-      <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="result" type="text" value={this.state.score.result}/> : this.state.score.result}</h3>
+       <div id="imgList">
+        <img src={this.state.score.url}/>
+       </div>
+        <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="happy" type="text" value={this.state.score.happy}/> : this.state.score.happy}</h3>
+        <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="mad" type="text" value={this.state.score.mad}/> : this.state.score.mad}</h3>
+        <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="url" type="text" value={this.state.score.url}/> : this.state.score.url}</h3>
+        <h3>{this.state.isEditing ? <input onChange={this.handleChange} name="result" type="text" value={this.state.score.result}/> : this.state.score.result}</h3>
 
       <label> {this.state.isEditing ?
-
        <select value={this.state.score.user_id} onChange={this.handleChange} name="user_id" type="num">
         <option value='1'>Julia</option>
         <option value='2'>Ann</option>
@@ -106,11 +105,10 @@ class SingleScore extends Component {
         <option value='7'>Tom</option>
        </select> : this.state.score.name}
       </label>
-
         </div>
         {this.state.isEditing ? <button type="submit">Save</button> : ''}
         {this.state.isEditing ? <button onClick={this.handleCancel}>Cancel</button> : ''}
-        </form>
+     </form>
         </div>
       );
     } else return <Loading />;
