+++ src/components/partials/Score.js
